@@ -9,23 +9,27 @@
   "prefer-stable": true,
   "minimum-stability": "dev",
   "require": {
-    "php": ">=5.6.0",
+    "php": ">=7.4",
     "ext-pdo": "*",
-    "byjg/anydataset": "4.0.*",
-    "byjg/anydataset-array": "4.0.*",
-    "psr/cache": "1.0.*",
-    "byjg/uri": "2.1.*"
+    "byjg/anydataset-array": "4.9.*",
+    "psr/cache": "^1.0",
+    "psr/log": "^1.0",
+    "byjg/uri": "4.9.*",
+    "ext-json": "*"
   },
   "require-dev": {
     "phpunit/phpunit": "5.7.*|7.4.*|^9.5",
-    "byjg/cache-engine": "4.0.*"
+    "byjg/cache-engine": "4.9.*"
   },
   "suggest": {
     "ext-curl": "*",
     "ext-pdo_mysql": "*",
     "ext-pdo_pgsql": "*",
     "ext-oci8": "*",
-    "byjg/cache-engine": "4.0.*"
+    "byjg/cache-engine": "4.9.*"
+  },
+  "provide": {
+    "byjg/anydataset-implementation": "1.0"
   },
   "license": "MIT"
 }
