@@ -5,6 +5,7 @@ namespace ByJG\AnyDataset\Db;
 use ByJG\AnyDataset\Core\Exception\IteratorException;
 use ByJG\AnyDataset\Core\GenericIterator;
 use ByJG\AnyDataset\Core\Row;
+use ByJG\Serializer\Exception\InvalidArgumentException;
 use PDO;
 use PDOStatement;
 
@@ -33,6 +34,7 @@ class DbIterator extends GenericIterator
     /**
      * @return int
      */
+    #[\ReturnTypeWillChange]
     public function count()
     {
         return $this->statement->rowCount();
@@ -40,7 +42,7 @@ class DbIterator extends GenericIterator
 
     /**
      * @return bool
-     * @throws \ByJG\Serializer\Exception\InvalidArgumentException
+     * @throws InvalidArgumentException
      */
     public function hasNext()
     {
@@ -54,18 +56,9 @@ class DbIterator extends GenericIterator
 
         $rowArray = $this->statement->fetch(PDO::FETCH_ASSOC);
         if (!empty($rowArray)) {
-            foreach ($rowArray as $key => $value) {
-                if (is_object($value)) {
-                    $rowArray[$key] = "[OBJECT]";
-                } else {
-                    $rowArray[$key] = $value;
-                }
-            }
             $singleRow = new Row($rowArray);
 
-            // Enfileira o registo
-            array_push($this->rowBuffer, $singleRow);
-            // Traz novos até encher o Buffer
+            $this->rowBuffer[] = $singleRow;
             if (count($this->rowBuffer) < DbIterator::RECORD_BUFFER) {
                 $this->hasNext();
             }
@@ -82,7 +75,7 @@ class DbIterator extends GenericIterator
     /**
      * @return Row
      * @throws IteratorException
-     * @throws \ByJG\Serializer\Exception\InvalidArgumentException
+     * @throws InvalidArgumentException
      */
     public function moveNext()
     {
