@@ -2,8 +2,9 @@
 
 namespace ByJG\AnyDataset\Db\Helpers;
 
-use ByJG\AnyDataset\Db\DbDriverInterface;
 use ByJG\AnyDataset\Core\Exception\NotAvailableException;
+use ByJG\AnyDataset\Db\DbDriverInterface;
+use ByJG\AnyDataset\Db\IsolationLevelEnum;
 
 class DbDblibFunctions extends DbBaseFunctions
 {
@@ -144,5 +145,21 @@ class DbDblibFunctions extends DbBaseFunctions
     public function hasForUpdate()
     {
         return false;
+    }
+
+    public function getIsolationLevelCommand($isolationLevel)
+    {
+        switch ($isolationLevel) {
+            case IsolationLevelEnum::READ_UNCOMMITTED:
+                return "SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED";
+            case IsolationLevelEnum::READ_COMMITTED:
+                return "SET TRANSACTION ISOLATION LEVEL READ COMMITTED";
+            case IsolationLevelEnum::REPEATABLE_READ:
+                return "SET TRANSACTION ISOLATION LEVEL REPEATABLE READ";
+            case IsolationLevelEnum::SERIALIZABLE:
+                return "SET TRANSACTION ISOLATION LEVEL SERIALIZABLE";
+            default:
+                return "";
+        }
     }
 }
