@@ -2,11 +2,17 @@
 
 namespace ByJG\AnyDataset\Db;
 
+use ByJG\AnyDataset\Core\Exception\NotAvailableException;
 use ByJG\Util\Uri;
 use PDO;
 
 class PdoMysql extends DbPdoDriver
 {
+
+    public static function schema()
+    {
+        return ['mysql', 'mariadb'];
+    }
 
     protected $mysqlAttr = [
         "ca" => PDO::MYSQL_ATTR_SSL_CA,
@@ -21,7 +27,7 @@ class PdoMysql extends DbPdoDriver
      * PdoMysql constructor.
      *
      * @param \ByJG\Util\Uri $connUri
-     * @throws \ByJG\AnyDataset\Core\Exception\NotAvailableException
+     * @throws NotAvailableException
      */
     public function __construct(Uri $connUri)
     {
@@ -31,7 +37,6 @@ class PdoMysql extends DbPdoDriver
 
         $postOptions = [
             PDO::MYSQL_ATTR_USE_BUFFERED_QUERY => true,
-            PDO::ATTR_EMULATE_PREPARES => true
         ];
 
         if (!empty($connUri->getQuery())) {
