@@ -2,16 +2,22 @@
 
 namespace ByJG\AnyDataset\Db;
 
+use ByJG\AnyDataset\Core\Exception\NotAvailableException;
 use ByJG\Util\Uri;
 
 class PdoOdbc extends DbPdoDriver
 {
 
+    public static function schema()
+    {
+        return ['odbc'];
+    }
+
     /**
      * PdoOdbc constructor.
      *
      * @param \ByJG\Util\Uri $connUri
-     * @throws \ByJG\AnyDataset\Core\Exception\NotAvailableException
+     * @throws NotAvailableException
      */
     public function __construct(Uri $connUri)
     {
