@@ -3,10 +3,14 @@
 namespace TestsDb\AnyDataset;
 
 use ByJG\AnyDataset\Core\Exception\NotImplementedException;
-use ByJG\AnyDataset\Db\DbCached;
 use ByJG\AnyDataset\Db\DbDriverInterface;
 use ByJG\AnyDataset\Db\DbPdoDriver;
+use ByJG\AnyDataset\Db\Exception\DbDriverNotConnected;
+use ByJG\AnyDataset\Db\Exception\TransactionNotStartedException;
+use ByJG\AnyDataset\Db\Exception\TransactionStartedException;
 use ByJG\AnyDataset\Db\Factory;
+use ByJG\AnyDataset\Db\IsolationLevelEnum;
+use ByJG\Cache\Psr16\ArrayCacheEngine;
 use PHPUnit\Framework\TestCase;
 
 abstract class BasePdo extends TestCase
@@ -24,7 +28,7 @@ abstract class BasePdo extends TestCase
      */
     public function setUp(): void
     {
-        $this->createInstance();
+        $this->dbDriver = $this->createInstance();
         $this->createDatabase();
         $this->populateData();
     }
@@ -40,7 +44,7 @@ abstract class BasePdo extends TestCase
         $array = $this->allData();
         foreach ($array as $param) {
             $this->dbDriver->execute(
-                "INSERT INTO Dogs (Breed, Name, Age) VALUES (:breed, :name, :age);",
+                "INSERT INTO Dogs (Breed, Name, Age, Weight) VALUES (:breed, :name, :age, :weight);",
                 $param
             );
         }
@@ -59,6 +63,7 @@ abstract class BasePdo extends TestCase
 
     public function tearDown(): void
     {
+        $this->dbDriver->reconnect();
         $this->deleteDatabase();
     }
 
@@ -69,19 +74,22 @@ abstract class BasePdo extends TestCase
                 'breed' => 'Mutt',
                 'name' => 'Spyke',
                 'age' => 8,
-                'id' => 1
+                'id' => 1,
+                'weight' =>  8.5
             ],
             [
                 'breed' => 'Brazilian Terrier',
                 'name' => 'Sandy',
                 'age' => 3,
-                'id' => 2
+                'id' => 2,
+                'weight' =>  3.8
             ],
             [
                 'breed' => 'Pincher',
                 'name' => 'Lola',
                 'age' => 1,
-                'id' => 3
+                'id' => 3,
+                'weight' =>  1.2
             ]
         ];
     }
@@ -128,7 +136,8 @@ abstract class BasePdo extends TestCase
                 'id',
                 'breed',
                 'name',
-                'age'
+                'age',
+                'weight'
             ],
             $allFields
         );
@@ -151,11 +160,10 @@ abstract class BasePdo extends TestCase
     {
         if (!$this->dbDriver->isSupportMultRowset()) {
             $this->markTestSkipped('Skipped: This DbDriver does not support multiple row set');
-            return;
         }
 
-        $sql = "INSERT INTO Dogs (Breed, Name, Age) VALUES ('Cat', 'Doris', 7); " .
-            "INSERT INTO Dogs (Breed, Name, Age) VALUES ('Dog', 'Lolla', 1); ";
+        $sql = "INSERT INTO Dogs (Breed, Name, Age, Weight) VALUES ('Cat', 'Doris', 7, 4.2); " .
+            "INSERT INTO Dogs (Breed, Name, Age, Weight) VALUES ('Dog', 'Lolla', 1, 1.4); ";
 
         $idInserted = $this->dbDriver->executeAndGetId($sql);
 
@@ -186,16 +194,17 @@ abstract class BasePdo extends TestCase
     public function testInsertSpecialChars()
     {
         $this->dbDriver->execute(
-            "INSERT INTO Dogs (Breed, Name, Age) VALUES ('Dog', '€ Sign Pètit Pannô', 6);"
+            "INSERT INTO Dogs (Breed, Name, Age, Weight) VALUES ('Dog', '€ Sign Pètit Pannô', 6, 3.2);"
         );
 
-        $iterator = $this->dbDriver->getIterator('select Id, Breed, Name, Age from Dogs where id = 4');
+        $iterator = $this->dbDriver->getIterator('select Id, Breed, Name, Age, Weight from Dogs where id = 4');
         $row = $iterator->toArray();
 
         $this->assertEquals(4, $row[0]["id"]);
         $this->assertEquals('Dog', $row[0]["breed"]);
         $this->assertEquals('€ Sign Pètit Pannô', $row[0]["name"]);
         $this->assertEquals(6, $row[0]["age"]);
+        $this->assertEquals(3.2, $row[0]["weight"]);
     }
 
     public function testEscapeQuote()
@@ -203,7 +212,7 @@ abstract class BasePdo extends TestCase
         $escapeQuote = $this->escapeQuote;
 
         $this->dbDriver->execute(
-            "INSERT INTO Dogs (Breed, Name, Age) VALUES ('Dog', 'Puppy${escapeQuote}s Master', 6);"
+            "INSERT INTO Dogs (Breed, Name, Age) VALUES ('Dog', 'Puppy{$escapeQuote}s Master', 6);"
         );
 
         $iterator = $this->dbDriver->getIterator('select Id, Breed, Name, Age from Dogs where id = 4');
@@ -240,7 +249,7 @@ abstract class BasePdo extends TestCase
         $escapeQuote = $this->escapeQuote;
 
         $this->dbDriver->execute(
-            "INSERT INTO Dogs (Breed, Name, Age) VALUES (:breed, 'Puppy${escapeQuote}s Master', :age);",
+            "INSERT INTO Dogs (Breed, Name, Age) VALUES (:breed, 'Puppy{$escapeQuote}s Master', :age);",
             [
                 "breed" => 'Dog',
                 "age" => 6
@@ -293,25 +302,33 @@ abstract class BasePdo extends TestCase
 
     public function testCachedResults()
     {
-        $dbCached = new DbCached($this->dbDriver, \ByJG\Cache\Factory::createArrayPool(), 600);
-
-        // Get the first from Db and then cache it;
-        $iterator = $dbCached->getIterator('select * from Dogs where id = :id', ['id' => 1]);
+        // Check with no cache at all
+        $iterator = $this->dbDriver->getIterator('select * from Dogs where id = :id', ['id' => 1]);
         $this->assertEquals(
             [
-                [ 'id'=> 1, 'breed' => "Mutt", 'name' => 'Spyke', "age" => 8, "__id" => 0, "__key" => 0],
+                [ 'id'=> 1, 'breed' => "Mutt", 'name' => 'Spyke', "age" => 8, "weight" => 8.5],
+            ],
+            $iterator->toArray()
+        );
+
+        $cacheEngine = new ArrayCacheEngine();
+        // Get the first from Db and then cache it;
+        $iterator = $this->dbDriver->getIterator('select * from Dogs where id = :id', ['id' => 1], $cacheEngine, 60);
+        $this->assertEquals(
+            [
+                [ 'id'=> 1, 'breed' => "Mutt", 'name' => 'Spyke', "age" => 8, "weight" => 8.5, "__id" => 0, "__key" => 0],
             ],
             $iterator->toArray()
         );
 
         // Remove it from DB (Still in cache) - Execute don't use cache
-        $dbCached->execute("delete from Dogs where id = :id", ['id' => 1]);
+        $this->dbDriver->execute("delete from Dogs where id = :id", ['id' => 1]);
 
         // Try get from cache
-        $iterator = $dbCached->getIterator('select * from Dogs where id = :id', ['id' => 1]);
+        $iterator = $this->dbDriver->getIterator('select * from Dogs where id = :id', ['id' => 1], $cacheEngine, 60);
         $this->assertEquals(
             [
-                [ 'id'=> 1, 'breed' => "Mutt", 'name' => 'Spyke', "age" => 8, "__id" => 0, "__key" => 0],
+                [ 'id'=> 1, 'breed' => "Mutt", 'name' => 'Spyke', "age" => 8, "weight" => 8.5, "__id" => 0, "__key" => 0],
             ],
             $iterator->toArray()
         );
@@ -319,34 +336,34 @@ abstract class BasePdo extends TestCase
 
     public function testCachedResultsNotFound()
     {
-        $dbCached = new DbCached($this->dbDriver, \ByJG\Cache\Factory::createArrayPool(), 600);
+        $cacheEngine = new ArrayCacheEngine();
 
         // Get the first from Db and then cache it;
-        $iterator = $dbCached->getIterator('select * from Dogs where id = :id', ['id' => 4]);
+        $iterator = $this->dbDriver->getIterator('select * from Dogs where id = :id', ['id' => 4], $cacheEngine, 60);
         $this->assertEquals(
             [],
             $iterator->toArray()
         );
-        $iterator = $dbCached->getIterator('select * from Dogs where id = :id', ['id' => 1]);
+        $iterator = $this->dbDriver->getIterator('select * from Dogs where id = :id', ['id' => 1], $cacheEngine, 60);
         $this->assertEquals(
             [
-                [ 'id'=> 1, 'breed' => "Mutt", 'name' => 'Spyke', "age" => 8, "__id" => 0, "__key" => 0],
+                [ 'id'=> 1, 'breed' => "Mutt", 'name' => 'Spyke', "age" => 8, "weight" => 8.5, "__id" => 0, "__key" => 0],
             ],
             $iterator->toArray()
         );
 
-        // Remove it from DB (Still in cache)
+        // Update Record
         $this->dbDriver->execute("INSERT INTO Dogs (Breed, Name, Age) VALUES (:breed, :name, :age);", ["breed" => "Cat", "name" => "Doris", "age" => 6]);
 
-        // Try get from cache
-        $iterator = $dbCached->getIterator('select * from Dogs where id = :id', ['id' => 1]);
+        // Try get from cache (should have the same result from before)
+        $iterator = $this->dbDriver->getIterator('select * from Dogs where id = :id', ['id' => 1], $cacheEngine, 60);
         $this->assertEquals(
             [
-                [ 'id'=> 1, 'breed' => "Mutt", 'name' => 'Spyke', "age" => 8, "__id" => 0, "__key" => 0],
+                [ 'id'=> 1, 'breed' => "Mutt", 'name' => 'Spyke', "age" => 8, "weight" => 8.5, "__id" => 0, "__key" => 0],
             ],
             $iterator->toArray()
         );
-        $iterator = $dbCached->getIterator('select * from Dogs where id = :id', ['id' => 4]);
+        $iterator = $this->dbDriver->getIterator('select * from Dogs where id = :id', ['id' => 4], $cacheEngine, 60);
         $this->assertEquals(
             [],
             $iterator->toArray()
@@ -357,5 +374,237 @@ abstract class BasePdo extends TestCase
         throw new NotImplementedException("Needs to be implemented for each database");
     }
 
+    public function testGetMetadata()
+    {
+        $metadata = $this->dbDriver->getDbHelper()->getTableMetadata($this->dbDriver, 'Dogs');
+
+        foreach ($metadata as $key => $field) {
+            unset($metadata[$key]['dbType']);
+        }
+
+        $this->assertEquals([
+            'id' => [
+                'name' => 'Id',
+                'required' => true,
+                'default' => null,
+                'phpType' => 'integer',
+                'length' => null,
+                'precision' => null,
+            ],
+            'breed' => [
+                'name' => 'Breed',
+                'required' => false,
+                'default' => null,
+                'phpType' => 'string',
+                'length' => 50,
+                'precision' => null,
+            ],
+            'name' => [
+                'name' => 'Name',
+                'required' => false,
+                'default' => null,
+                'phpType' => 'string',
+                'length' => 50,
+                'precision' => null,
+            ],
+            'age' => [
+                'name' => 'Age',
+                'required' => false,
+                'default' => null,
+                'phpType' => 'integer',
+                'length' => null,
+                'precision' => null,
+            ],
+            'weight' => [
+                'name' => 'Weight',
+                'required' => false,
+                'default' => null,
+                'phpType' => 'float',
+                'length' => 10,
+                'precision' => 2,
+            ],
+        ], $metadata);
+    }
+
+    public function testDisconnect()
+    {
+        $iterator = $this->dbDriver->getIterator('select Id, Breed, Name, Age from Dogs where id = 1');
+        $row = $iterator->toArray();
+        $this->assertEquals(1, $row[0]["id"]);
+
+        $this->dbDriver->disconnect();
+
+        $this->expectException(DbDriverNotConnected::class);
+        $iterator = $this->dbDriver->getIterator('select Id, Breed, Name, Age from Dogs where id = 1');
+    }
+
+    public function testReconnect()
+    {
+        $this->assertFalse($this->dbDriver->reconnect());
+        $this->assertTrue($this->dbDriver->reconnect(true));
+        $iterator = $this->dbDriver->getIterator('select Id, Breed, Name, Age from Dogs where id = 1');
+    }
+
+    public function testCommitTransaction()
+    {
+        $this->assertFalse($this->dbDriver->hasActiveTransaction());
+        $this->assertNull($this->dbDriver->activeIsolationLevel());
+        $this->dbDriver->beginTransaction(IsolationLevelEnum::SERIALIZABLE);
+        $this->assertTrue($this->dbDriver->hasActiveTransaction());
+        $this->assertEquals(IsolationLevelEnum::SERIALIZABLE, $this->dbDriver->activeIsolationLevel());
+
+        $idInserted = $this->dbDriver->executeAndGetId(
+            "INSERT INTO Dogs (Breed, Name, Age) VALUES ('Cat', 'Doris', 7);"
+        );
+        $this->dbDriver->commitTransaction();
+        $this->assertFalse($this->dbDriver->hasActiveTransaction());
+        $this->assertNull($this->dbDriver->activeIsolationLevel());
+
+        $this->assertEquals(4, $idInserted);
+
+        $iterator = $this->dbDriver->getIterator('select Id, Breed, Name, Age from Dogs where id = 4');
+        $row = $iterator->toArray();
+
+        $this->assertEquals(4, $row[0]["id"]);
+        $this->assertEquals('Cat', $row[0]["breed"]);
+        $this->assertEquals('Doris', $row[0]["name"]);
+        $this->assertEquals(7, $row[0]["age"]);
+    }
+
+    public function testRollbackTransaction()
+    {
+        $this->assertFalse($this->dbDriver->hasActiveTransaction());
+        $this->assertNull($this->dbDriver->activeIsolationLevel());
+        $this->dbDriver->beginTransaction(IsolationLevelEnum::REPEATABLE_READ);
+        $this->assertTrue($this->dbDriver->hasActiveTransaction());
+        $this->assertEquals(IsolationLevelEnum::REPEATABLE_READ, $this->dbDriver->activeIsolationLevel());
+
+        $idInserted = $this->dbDriver->executeAndGetId(
+            "INSERT INTO Dogs (Breed, Name, Age) VALUES ('Cat', 'Doris', 7);"
+        );
+        $this->dbDriver->rollbackTransaction();
+        $this->assertFalse($this->dbDriver->hasActiveTransaction());
+        $this->assertNull($this->dbDriver->activeIsolationLevel());
+
+        $this->assertEquals(4, $idInserted);
+
+        $iterator = $this->dbDriver->getIterator('select Id, Breed, Name, Age from Dogs where id = 4');
+        $row = $iterator->toArray();
+
+        $this->assertEmpty($row);
+    }
+
+    public function testCommitWithoutTransaction()
+    {
+        $this->expectException(TransactionNotStartedException::class);
+        $this->dbDriver->commitTransaction();
+    }
+
+    public function testRollbackWithoutTransaction()
+    {
+        $this->expectException(TransactionNotStartedException::class);
+        $this->dbDriver->rollbackTransaction();
+    }
+
+    public function testRequiresTransaction()
+    {
+        $this->assertFalse($this->dbDriver->hasActiveTransaction());
+        $this->assertNull($this->dbDriver->activeIsolationLevel());
+        $this->assertEquals(0, $this->dbDriver->remainingCommits());
+
+        $this->dbDriver->beginTransaction(IsolationLevelEnum::READ_COMMITTED);
+        $this->assertTrue($this->dbDriver->hasActiveTransaction());
+        $this->assertEquals(IsolationLevelEnum::READ_COMMITTED, $this->dbDriver->activeIsolationLevel());
+        $this->assertEquals(1, $this->dbDriver->remainingCommits());
+
+        $this->dbDriver->requiresTransaction();
+        $this->assertTrue($this->dbDriver->hasActiveTransaction());
+        $this->assertEquals(IsolationLevelEnum::READ_COMMITTED, $this->dbDriver->activeIsolationLevel());
+        $this->assertEquals(1, $this->dbDriver->remainingCommits());
+
+        $this->dbDriver->commitTransaction();
+        $this->assertFalse($this->dbDriver->hasActiveTransaction());
+        $this->assertNull($this->dbDriver->activeIsolationLevel());
+        $this->assertEquals(0, $this->dbDriver->remainingCommits());
+    }
+
+    public function testRequiresTransactionWithoutTransaction()
+    {
+        $this->expectException(TransactionNotStartedException::class);
+        $this->dbDriver->requiresTransaction();
+    }
+
+    public function testBeginTransactionTwice()
+    {
+        $this->dbDriver->beginTransaction(IsolationLevelEnum::READ_UNCOMMITTED);
+        $this->expectException(TransactionStartedException::class);
+        try {
+            $this->dbDriver->beginTransaction();
+        } finally {
+            $this->dbDriver->rollbackTransaction();
+            $this->assertFalse($this->dbDriver->hasActiveTransaction());
+            $this->assertNull($this->dbDriver->activeIsolationLevel());
+            $this->assertEquals(0, $this->dbDriver->remainingCommits());
+        }
+    }
+
+    public function testJoinTransaction()
+    {
+        $this->dbDriver->beginTransaction(IsolationLevelEnum::READ_UNCOMMITTED);
+        $this->assertEquals(1, $this->dbDriver->remainingCommits());
+        $this->assertEquals(IsolationLevelEnum::READ_UNCOMMITTED, $this->dbDriver->activeIsolationLevel());
+
+        $this->dbDriver->beginTransaction(IsolationLevelEnum::READ_UNCOMMITTED, true);
+        $this->assertEquals(2, $this->dbDriver->remainingCommits());
+        $this->assertEquals(IsolationLevelEnum::READ_UNCOMMITTED, $this->dbDriver->activeIsolationLevel());
+
+        $this->dbDriver->commitTransaction();
+        $this->assertEquals(1, $this->dbDriver->remainingCommits());
+        $this->assertTrue($this->dbDriver->hasActiveTransaction());
+        $this->assertEquals(IsolationLevelEnum::READ_UNCOMMITTED, $this->dbDriver->activeIsolationLevel());
+
+        $this->dbDriver->commitTransaction();
+        $this->assertEquals(0, $this->dbDriver->remainingCommits());
+        $this->assertFalse($this->dbDriver->hasActiveTransaction());
+        $this->assertNull($this->dbDriver->activeIsolationLevel());
+    }
+
+    public function testTwoDifferentTransactions()
+    {
+        $dbDriver1 = $this->createInstance();
+        $dbDriver2 = $this->createInstance();
+
+        // Make sure there is no record in the database
+        $iterator = $dbDriver1->getIterator('select Id, Breed, Name, Age from Dogs where id = 4');
+        $row = $iterator->toArray();
+        $this->assertEmpty($row);
+        $iterator = $dbDriver2->getIterator('select Id, Breed, Name, Age from Dogs where id = 4');
+        $row = $iterator->toArray();
+        $this->assertEmpty($row);
+
+        // Start a transaction on the first connection
+        $dbDriver1->beginTransaction(IsolationLevelEnum::SERIALIZABLE);
+        $idInserted = $dbDriver1->executeAndGetId(
+            "INSERT INTO Dogs (Breed, Name, Age) VALUES ('Cat', 'Doris', 7);"
+        );
+
+        // Check if the record is there
+        $iterator = $dbDriver1->getIterator('select Id, Breed, Name, Age from Dogs where id = 4');
+        $row = $iterator->toArray();
+        $this->assertNotEmpty($row);
+
+        // Check if the record is not there on the second connection (due to isolation level)
+        $iterator = $dbDriver2->getIterator('select Id, Breed, Name, Age from Dogs where id = 4');
+        $row = $iterator->toArray();
+        $this->assertEmpty($row);
+
+        // Commit the transaction
+        $dbDriver1->commitTransaction();
+
+        // Check if the second transaction can read
+        $iterator = $dbDriver2->getIterator('select Id, Breed, Name, Age from Dogs where id = 4');
+        $row = $iterator->toArray();
+        $this->assertNotEmpty($row);
+    }
 }
 
