@@ -23,18 +23,18 @@ class PdoPostgresTest extends BasePdo
             $password = "";
         }
 
-        $this->dbDriver = Factory::getDbRelationalInstance("pgsql://postgres:$password@$host");
-        $exists = $this->dbDriver->getScalar('select count(1) from pg_catalog.pg_database where datname = \'test\'');
+        $dbDriver = Factory::getDbRelationalInstance("pgsql://postgres:$password@$host");
+        $exists = $dbDriver->getScalar('select count(1) from pg_catalog.pg_database where datname = \'test\'');
         if ($exists == 0) {
-            $this->dbDriver->execute('CREATE DATABASE test');
+            $dbDriver->execute('CREATE DATABASE test');
         }
-        $this->dbDriver = Factory::getDbRelationalInstance("pgsql://postgres:$password@$host/test");
+        return Factory::getDbRelationalInstance("pgsql://postgres:$password@$host/test");
     }
 
     protected function createDatabase()
     {
         //create the database
-        $this->dbDriver->execute("CREATE TABLE Dogs (Id SERIAL PRIMARY KEY, Breed VARCHAR(50), Name VARCHAR(50), Age INTEGER)");
+        $this->dbDriver->execute("CREATE TABLE Dogs (Id SERIAL PRIMARY KEY, Breed VARCHAR(50), Name VARCHAR(50), Age INTEGER, Weight NUMERIC(10,2))");
     }
 
     public function deleteDatabase()
@@ -54,5 +54,57 @@ class PdoPostgresTest extends BasePdo
         $this->expectException(\PDOException::class);
         
         parent::testDontParseParam_3();
+    }
+
+    public function testGetMetadata()
+    {
+        $metadata = $this->dbDriver->getDbHelper()->getTableMetadata($this->dbDriver, 'Dogs');
+
+        foreach ($metadata as $key => $field) {
+            unset($metadata[$key]['dbType']);
+        }
+
+        $this->assertEquals([
+            'id' => [
+                'name' => 'id',
+                'required' => true,
+                'default' => "nextval('dogs_id_seq'::regclass)",
+                'phpType' => 'integer',
+                'length' => null,
+                'precision' => null,
+            ],
+            'breed' => [
+                'name' => 'breed',
+                'required' => false,
+                'default' => null,
+                'phpType' => 'string',
+                'length' => 50,
+                'precision' => null,
+            ],
+            'name' => [
+                'name' => 'name',
+                'required' => false,
+                'default' => null,
+                'phpType' => 'string',
+                'length' => 50,
+                'precision' => null,
+            ],
+            'age' => [
+                'name' => 'age',
+                'required' => false,
+                'default' => null,
+                'phpType' => 'integer',
+                'length' => null,
+                'precision' => null,
+            ],
+            'weight' => [
+                'name' => 'weight',
+                'required' => false,
+                'default' => null,
+                'phpType' => 'float',
+                'length' => 10,
+                'precision' => 2,
+            ],
+        ], $metadata);
     }
 }
